requirements.txt
asgiref==3.8.1
Django==5.0
django-autoslug==1.9.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==35.2.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
